Tiles now rotate with correct colors
Looks like I broke some of the keyboard structure.
When I shrink the window now the keys move around.
I thought 'resize: none' was preventing this, but
apparently not.  Making things position: relative
around things position: fixed is still mysterious.

Struggled to get a valid test for Tile to show the
rotateTile method is called.  Used depricated
method.  Struggled with Jest.spyOn for way too long.

fixed some CSS in Board/Tile

Added theWord to the store and a way to set it.
Updated logic to determin the correct status for
each letter in the submitted valid word.