Stop adding letters when secret word is guessed.
Added setting the word of the day and a status
of solved: boolean.  Currently solved == true
stops the entry of any more letters.  It will
also be used to trigger some kind of notice
that the game is over and the player won.

I have also broken some tests... A problem for
another day.